refactor(mini chart) Move tooltip to area component
