+++ src/ui/MiniChart/Area.svelte
@@ -1,6 +1,9 @@
 <script lang="ts">
   import type { Props } from './svelte'
+
   import Chart from './index.svelte'
+  import { getTooltipContext } from './tooltip'
+  import { getValue } from './utils'
 
   let className = ''
   export { className as class }
@@ -11,6 +14,22 @@
   export let valueKey: Props['valueKey'] = undefined
   export let style: Props['style'] = undefined
 
+  export let tooltipVisible = false
+  export let tooltipSyncKey: string | undefined = undefined
+  export let formatTooltipValue = (value: number) => value.toString()
+
+  const { offsetMap, getOffset, updateOffset } = getTooltipContext()
+
+  $: offset = getOffset($offsetMap, tooltipSyncKey)
+  $: currentValue = getValueAt(offset, width)
+  $: valueFormatted = currentValue !== undefined ? formatTooltipValue(currentValue) : currentValue
+
+  function getValueAt(offset: number | null, width: number) {
+    if (!offset) return
+    const valueIndex = Math.round((offset / width) * (data.length - 1))
+    return getValue(data[valueIndex])
+  }
+
   export function getAreaPoints(points: Props['points'], linePoints: string) {
     const [startX, startY] = points[0].split(',')
     const [lastX] = points[points.length - 1].split(',')
@@ -19,7 +38,18 @@
   }
 </script>
 
-<Chart {data} {width} {height} {valueKey} class={className} {style} let:points let:linePoints>
+<Chart
+  {data}
+  {width}
+  {height}
+  {valueKey}
+  {style}
+  class="relative {className}"
+  on:mousemove={({ offsetX }) => updateOffset(offsetX, tooltipSyncKey)}
+  on:mouseleave={() => updateOffset(null, tooltipSyncKey)}
+  let:points
+  let:linePoints
+>
   <polyline points={getAreaPoints(points, linePoints)} fill="url(#{id}-area)" />
   <defs>
     <linearGradient id="{id}-area" x1="0" x2="0" y1="0" y2="2">
@@ -27,4 +57,23 @@
       <stop offset="60%" stop-color="var(--white)" stop-opacity="0" />
     </linearGradient>
   </defs>
+
+  {#if tooltipVisible && offset}
+    <line x1={offset} x2={offset} y1="0" y2={height} stroke="var(--waterloo)" />
+
+    <foreignObject {width} {height}>
+      <tooltip class="c-fiord">{valueFormatted}</tooltip>
+    </foreignObject>
+  {/if}
 </Chart>
+
+<style>
+  tooltip {
+    position: absolute;
+    top: 0;
+    left: 0;
+    background: rgba(255, 255, 255, 0.7);
+    border-radius: 4px;
+    padding: 2px 4px;
+  }
+</style>
